requirements.txt
streamlit==1.27.*
requests==2.26.0
streamlit-webcam-example